# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing Telegram Bot Electro Tools settings and metrics",
        "title": "Telegram Bot Electro Tools API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:54321",
    "basePath": "/",
    "paths": {
        "/api/v1/metrics": {
            "get": {
                "description": "Get application metrics in Prometheus format",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Get metrics",
                "responses": {
                    "200": {
                        "description": "Metrics data",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current bot settings\nUpdate bot settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "settings",
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "Settings to update",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current bot settings\nUpdate bot settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "settings",
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "Settings to update",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.GetSettingsResponse": {
            "type": "object",
            "properties": {
                "site_url": {
                    "type": "string"
                },
                "trigger_message_count": {
                    "type": "integer"
                }
            }
        },
        "api.UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "site_url": {
                    "type": "string"
                },
                "trigger_message_count": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}